+++ craft_parts/executor/organize.py
@@ -32,10 +32,7 @@ from typing import TYPE_CHECKING
 
 from craft_parts import errors
 from craft_parts.utils import file_utils, path_utils
-from craft_parts.utils.partition_utils import (
-    BUILD_PARTITION,
-    DEFAULT_PARTITION,
-)
+from craft_parts.utils.partition_utils import DEFAULT_PARTITION
 
 if TYPE_CHECKING:
     from collections.abc import Mapping
@@ -69,24 +66,10 @@ def organize_files(  # noqa: PLR0912
         the default partition.
     """
     for key in sorted(file_map, key=lambda x: ["*" in x, x]):
-        src_partition_pair = path_utils.get_partition_and_path(key, default_partition)
-        src = get_src_path(
-            src_partition_pair, part_name, install_dir_map, default_partition
-        )
-
-        # Remove the leading slash so the path actually joins
-        # Also trailing slash is significant, be careful if using pathlib!
-        dst_partition_pair = path_utils.get_partition_and_path(
-            file_map[key].lstrip("/"), default_partition
-        )
+        src = get_src_path(key, part_name, install_dir_map, default_partition)
         dst, dst_string = get_dst_path(
-            dst_partition_pair, part_name, install_dir_map, default_partition
+            key, file_map, install_dir_map, default_partition
         )
-
-        # If the destinations ends with a slash character and it doesn't exist,
-        # create it as a directory.
-        if dst.endswith("/") and not os.path.exists(dst.rstrip("/")):  # noqa: PTH110
-            os.makedirs(dst)  # noqa: PTH103
 
         sources = iglob(src, recursive=True)  # noqa: PTH207
 
@@ -99,8 +82,7 @@ def organize_files(  # noqa: PLR0912
             # Organize a dir to a dir
             if os.path.isdir(src) and "*" not in key:  # noqa: PTH112
                 file_utils.link_or_copy_tree(src, dst)
-                if src_partition_pair.partition != BUILD_PARTITION:
-                    shutil.rmtree(src)
+                shutil.rmtree(src)
                 continue
 
             # Organize a "not dir" (file, character device, etc.) to a "not dir"
@@ -154,28 +136,23 @@ def organize_files(  # noqa: PLR0912
                     )
 
             os.makedirs(os.path.dirname(dst), exist_ok=True)  # noqa: PTH103, PTH120
-            if src_partition_pair.partition == BUILD_PARTITION:
-                if os.path.isdir(src):  # noqa: PTH112
-                    file_utils.link_or_copy_tree(src, dst)
-                else:
-                    file_utils.link_or_copy(src, dst)
-            else:
-                file_utils.move(src, dst)
+            file_utils.move(src, dst)
 
 
 def get_src_path(
-    src_partition_path: path_utils.PartitionPathPair,
+    key: str,
     part_name: str,
     install_dir_map: Mapping[str | None, Path],
     default_partition: str,
 ) -> str:
     """Return the full path for a relative source."""
-    src_partition, src_inner_path = src_partition_path
+    src_partition, src_inner_path = path_utils.get_partition_and_path(
+        key, default_partition
+    )
 
     if src_partition and src_partition not in [
         default_partition,
         DEFAULT_PARTITION,
-        BUILD_PARTITION,
     ]:
         raise errors.FileOrganizeError(
             part_name=part_name,
@@ -193,21 +170,18 @@ def get_src_path(
 
 
 def get_dst_path(
-    dst_partition_path: path_utils.PartitionPathPair,
-    part_name: str,
+    key: str,
+    file_map: dict[str, str],
     install_dir_map: Mapping[str | None, Path],
     default_partition: str,
 ) -> tuple[str, str]:
     """Return the full destination path and log-friendly representation of a destination."""
-    # The build pseudo-partition can only be specified in the left hand
-    # side of the organize item.
-    dst_partition, dst_inner_path = dst_partition_path
-
-    if dst_partition == BUILD_PARTITION:
-        raise errors.FileOrganizeError(
-            part_name=part_name,
-            message="Cannot organize files into the build directory",
-        )
+    # Remove the leading slash so the path actually joins
+    # Also trailing slash is significant, be careful if using pathlib!
+    dst_partition, dst_inner_path = path_utils.get_partition_and_path(
+        file_map[key].lstrip("/"),
+        default_partition,
+    )
 
     # Replace default partition default name with alias name to allow
     # using (default) in paths even with aliased default partition